mnist4 attempt
still not working, same problem: out of range subscript in sub2ind.  I
made nn03 to prove that it's not just the size of the matrix (this is
the same as nn01 except I changed the number of training examples).
nn02 doesn't work and it appears to be something associated with the
source data.